Implement UX switch between summarized and granular attendance data. Split component into separate file
