+++ frontend/components/meeting-feedback.jsx
@@ -2,45 +2,8 @@ import React, { useState } from 'react'
 
 import RatingInput from './manage/meeting-rating-input';
 import MeetingReflectionInput from './manage/meeting-reflection-input';
+import AttendanceInput from './manage/meeting-attendance-input';
 import DelayedPostForm from './manage/delayed-post-form';
-
-const GranularAttendanceInput = ({value, onChange, learners}) => {
-
-  var parsedValue = JSON.parse(value)
-
-  const handleCheck = userId => {
-    if (parsedValue[userId]) {
-      delete parsedValue[userId]
-    } else {
-      parsedValue[userId] = true
-    }
-    onChange({"granular_attendance": JSON.stringify(parsedValue)})
-  }
-
-  return (
-    <div id="div_id_granular_attendance" className="form-group">
-      <label htmlFor="id_granular_attendance" className="col-form-label">Who attended?</label> 
-      {
-        learners.map(user => (
-          <div className="form-check" key={user.id}>
-            <input className="form-check-input" type="checkbox" value="" id="checkChecked" checked={parsedValue[user.id]} onChange={e => handleCheck(user.id)}/>
-            <label className="form-check-label" htmlFor="checkChecked">
-              {`${user.name} <${user.email || user.mobile}>`}
-            </label>
-          </div>
-        ))
-      }
-    </div>
-  );
-}
-
-const AttendanceInput = ({value, onChange}) => 
-  <div id="div_id_attendance" className="form-group">
-    <label htmlFor="id_attendance" className="col-form-label">How many people attended?</label> 
-    <div>
-      <input type="number" name="attendance" min="0" className="numberinput form-control form-control form-control" value={value} onChange={e => onChange({'attendance': e.target.value})} id="id_attendance" /> 
-    </div> 
-  </div>;
 
 
 const MeetingFeedbackForm = props => {
@@ -48,12 +11,7 @@ const MeetingFeedbackForm = props => {
   return (
     <form>
       <RatingInput value={formData.rating} onChange={updateForm}/>
-      <GranularAttendanceInput 
-        value={formData.granular_attendance}
-        onChange={updateForm}
-        learners={window.learners}
-      />
-      <AttendanceInput value={formData.attendance} onChange={updateForm}/>
+      <AttendanceInput formData={formData} onChange={updateForm}/>
       <MeetingReflectionInput value={formData.reflection} onChange={updateForm} />
     </form>
   )
